refactor: move add_vertical_line_with_annotation_in_legend to plots.py
