+++ backend/protzilla/data_analysis/plots.py
@@ -5,6 +5,8 @@ import numpy as np
 import pandas as pd
 import plotly.express as px
 import plotly.graph_objects as go
+from plotly import graph_objects as go
+from plotly.graph_objs import Figure
 from scipy import stats
 from sklearn.metrics.pairwise import cosine_similarity, euclidean_distances
 
@@ -466,3 +468,31 @@ def prot_quant_plot(
     )
 
     return dict(plots=[fig])
+
+
+def add_vertical_line_with_annotation_in_legend(
+    fig: Figure, dash: str, annotation: str, x_value: float, color: str = "blue"
+) -> None:
+    """
+    Adds a vertical line to a Plotly figure and includes a corresponding entry in the legend
+    without displaying an additional visible trace in the plot.
+
+    :param fig: Plotly Figure object to which the vertical line and legend entry are added.
+    :param dash: Line style for the vertical line (e.g., "solid", "dash", "dot").
+    :param annotation: Text to display in the legend corresponding to the vertical line.
+    :param x_value: X-coordinate at which to draw the vertical line.
+    :param color: Color of the vertical line and legend entry (default is "blue").
+    :return: None
+    """
+    # add vertical line
+    fig.add_vline(x=x_value, line_color=color, line_dash=dash, line_width=2)
+    # add annotation of the line to the legend
+    fig.add_trace(
+        go.Scatter(
+            x=[None],
+            y=[None],
+            mode="lines",
+            name=annotation,
+            line=dict(color=color, width=2, dash=dash),
+        )
+    )
